+++ tests/conftest.py
@@ -42,8 +42,9 @@ except ImportError:
     has_egl = False
 
 if not has_display and has_egl:
-    # It is necessary to configure pyglet in headless mode if necessary before importing Genesis
-    pyglet.options["headless"] = True
+    # It is necessary to configure pyglet in headless mode if necessary before importing Genesis.
+    # Note that environment variables are used instead of global options to ease option propagation to subprocesses.
+    os.environ["PYGLET_HEADLESS"] = "1"
     os.environ["GS_VIEWER_ALLOW_OFFSCREEN"] = "1"
 
 IS_INTERACTIVE_VIEWER_AVAILABLE = has_display or has_egl
